requirements.txt
requests
python-dotenv
pandas
numpy
tqdm
discord-webhook
pytz
